Explicitly convert Long to Double when type casting
```
// Scala 2.12. This is fine
Long x = 100000000000000
Double y =  y

// Scala 2.13. It has to be written
Long x = 100000000000000
Double y = x.toDouble
```

Otherwise, we're getting errors like:
```
[error] /Users/ioanna_kokkini/Projects/frontend/common/app/common/metrics.scala:29:20: Widening conversion from Long to Double is deprecated because it loses precision. Write `.toDouble` instead.
[error]       totalGcCount - lastGcCount.getAndSet(totalGcCount)
[error]                    ^
```

+++ common/app/metrics/FrontendMetrics.scala
@@ -138,7 +138,7 @@ object DurationMetric {
   def withMetrics[A](metric: DurationMetric)(block: => A): A = {
     val stopWatch: StopWatch = new StopWatch
     val result = block
-    metric.recordDuration(stopWatch.elapsed)
+    metric.recordDuration(stopWatch.elapsed.toDouble)
     result
   }
 }
